Module rs9p::fcall

9P protocol data types and constants.

Protocol

9P2000.L

Modules

p92000
Old 9P2000 protocol types

Structs

Data
Data type used in Rread and Twrite

DirEntry
Directory entry used in Rreaddir

DirEntryData
Directory entry array

Flock
Similar to Linux struct flock

GetattrMask
Bits in mask and valid of Tgetattr and Rgetattr.

Getlock
Getlock structure

LockFlag
File lock flags, Flock.flags

LockStatus
File lock status

LockType
File lock type, Flock.typ

Msg
Envelope for 9P messages

Qid
Server side data type for path tracking

QidType
Bits in Qid.typ

SetAttr
Subset of Stat used for Tsetattr

SetattrMask
Bits in mask of Tsetattr.

Stat
File attributes corresponding to struct stat of Linux.

Statfs
Filesystem information corresponding to struct statfs of Linux.

Time
Time struct

Enums

Fcall
A data type encapsulating the various 9P messages

MsgType
Message type, 9P operations

Constants

IOHDRSZ
Ample room for Twrite/Rread header

NOFID
Special value which Tattach with no auth must use as afid

NONUNAME
Special uid which Tauth/Tattach use as n_uname to indicate no uid is specified

NOTAG
Special tag which Tversion/Rversion must use as tag

P92000
9P2000 version string

P92000L
9P2000.L version string

READDIRHDRSZ
Room for readdir header

V9FS_PORT
v9fs default port

VERSION_UNKNOWN
The version string that comes with Rversion when the server does not understand the client’s version string