Module rs9p::serialize

Serialize/deserialize 9P messages into/from binary.

Structs

Decoder
A wrapper class of ReadBytesExt to provide operator overloads for deserializing

Encoder
A wrapper class of WriteBytesExt to provide operator overloads for serializing

SResult
A serializing specific result to overload operators on Result

Traits

Decodable
Trait representing a type which can be deserialized from binary

Encodable
Trait representing a type which can be serialized into binary

Functions

read_msg
Helper function to read a 9P message from a byte-oriented stream

write_msg
Helper function to write a 9P message into a byte-oriented stream