Trait rs9p::srv::Filesystem

pub trait Filesystem: Send {
    type Fid: Send + Sync + Default;

    #[must_use]
    fn rstatfs<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rlopen<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _flags: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rlcreate<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _name: &'life2 str, 
        _flags: u32, 
        _mode: u32, 
        _gid: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rsymlink<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _name: &'life2 str, 
        _sym: &'life3 str, 
        _gid: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rmknod<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _name: &'life2 str, 
        _mode: u32, 
        _major: u32, 
        _minor: u32, 
        _gid: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rrename<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        __arg2: &'life2 Fid<Self::Fid>, 
        _name: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rreadlink<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rgetattr<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _req_mask: GetattrMask
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rsetattr<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _valid: SetattrMask, 
        _stat: &'life2 SetAttr
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rxattrwalk<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        __arg2: &'life2 Fid<Self::Fid>, 
        _name: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rxattrcreate<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _name: &'life2 str, 
        _attr_size: u64, 
        _flags: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rreaddir<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _offset: u64, 
        _count: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rfsync<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rlock<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _lock: &'life2 Flock
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rgetlock<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _lock: &'life2 Getlock
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rlink<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        __arg2: &'life2 Fid<Self::Fid>, 
        _name: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rmkdir<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _name: &'life2 str, 
        _mode: u32, 
        _gid: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rrenameat<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _oldname: &'life2 str, 
        __arg3: &'life3 Fid<Self::Fid>, 
        _newname: &'life4 str
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn runlinkat<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _name: &'life2 str, 
        _flags: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rauth<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _uname: &'life2 str, 
        _aname: &'life3 str, 
        _n_uname: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rattach<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _afid: Option<&'life2 Fid<Self::Fid>>, 
        _uname: &'life3 str, 
        _aname: &'life4 str, 
        _n_uname: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rflush<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        _old: Option<&'life1 Fcall>
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rwalk<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _new: &'life2 Fid<Self::Fid>, 
        _wnames: &'life3 [String]
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rread<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _offset: u64, 
        _count: u32
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rwrite<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>, 
        _offset: u64, 
        _data: &'life2 Data
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rclunk<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rremove<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 Fid<Self::Fid>
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    #[must_use]
    fn rversion<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        msize: u32, 
        ver: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

}

Filesystem server trait.

Implementors can represent an error condition by returning an Err. Otherwise, they must return Fcall with the required fields filled.

The default implementation, returning EOPNOTSUPP error, is provided to the all methods except Rversion. The default implementation of Rversion returns a message accepting 9P2000.L.

NOTE

Defined as Srv in 9p.h of Plan 9.

Protocol

9P2000.L

Associated Types

type Fid: Send + Sync + Default

User defined fid type to be associated with a client’s fid.

Loading content...
Provided methods

#[must_use]
fn rstatfs<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rlopen<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_flags: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rlcreate<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_name: &'life2 str,
_flags: u32,
_mode: u32,
_gid: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rsymlink<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_name: &'life2 str,
_sym: &'life3 str,
_gid: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rmknod<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_name: &'life2 str,
_mode: u32,
_major: u32,
_minor: u32,
_gid: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rrename<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
__arg2: &'life2 Fid<Self::Fid>,
_name: &'life3 str
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rreadlink<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rgetattr<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_req_mask: GetattrMask
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rsetattr<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_valid: SetattrMask,
_stat: &'life2 SetAttr
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rxattrwalk<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
__arg2: &'life2 Fid<Self::Fid>,
_name: &'life3 str
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rxattrcreate<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_name: &'life2 str,
_attr_size: u64,
_flags: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rreaddir<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_offset: u64,
_count: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rfsync<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rlock<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_lock: &'life2 Flock
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rgetlock<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_lock: &'life2 Getlock
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rlink<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
__arg2: &'life2 Fid<Self::Fid>,
_name: &'life3 str
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rmkdir<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_name: &'life2 str,
_mode: u32,
_gid: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rrenameat<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_oldname: &'life2 str,
__arg3: &'life3 Fid<Self::Fid>,
_newname: &'life4 str
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn runlinkat<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_name: &'life2 str,
_flags: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rauth<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_uname: &'life2 str,
_aname: &'life3 str,
_n_uname: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rattach<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_afid: Option<&'life2 Fid<Self::Fid>>,
_uname: &'life3 str,
_aname: &'life4 str,
_n_uname: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rflush<'life0, 'life1, 'async_trait>(
&'life0 self,
_old: Option<&'life1 Fcall>
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rwalk<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_new: &'life2 Fid<Self::Fid>,
_wnames: &'life3 [String]
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rread<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_offset: u64,
_count: u32
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rwrite<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>,
_offset: u64,
_data: &'life2 Data
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rclunk<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rremove<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 Fid<Self::Fid>
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

#[must_use]
fn rversion<'life0, 'life1, 'async_trait>(
&'life0 self,
msize: u32,
ver: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Fcall>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Loading content...