Module rs9p::fcall

9P protocol data types and constants.

Protocol

9P2000.L

Reexports

pub use self::ltype::LockType;

pub use self::lflag::LockFlag;

pub use self::lstatus::LockStatus;

pub use self::qt::QidType;

pub use self::getattr::GetattrMask;

pub use self::setattr::SetattrMask;

Modules

getattr
Bits in mask and valid of Tgetattr and Rgetattr.

lflag
File lock flags, Flock.flags

lstatus
File lock status

ltype
File lock type, Flock.typ

p92000
Old 9P2000 protocol types

qt
Bits in Qid.typ

setattr
Bits in mask of Tsetattr.

Structs

Data
Data type used in Rread and Twrite

DirEntry
Directory entry used in Rreaddir

DirEntryData
Directory entry array

Flock
Similar to Linux struct flock

Getlock
Getlock structure

Msg
Envelope for 9P messages

Qid
Server side data type for path tracking

SetAttr
Subset of Stat used for Tsetattr

Stat
File attributes corresponding to struct stat of Linux.

Statfs
Filesystem information corresponding to struct statfs of Linux.

Time
Time struct

Enums

Fcall
A data type encapsulating the various 9P messages

MsgType
Message type, 9P operations

Constants

IOHDRSZ
Ample room for Twrite/Rread header

NOFID
Special value which Tattach with no auth must use as afid

NONUNAME
Special uid which Tauth/Tattach use as n_uname to indicate no uid is specified

NOTAG
Special tag which Tversion/Rversion must use as tag

P92000
9P2000 version string

P92000L
9P2000.L version string

READDIRHDRSZ
Room for readdir header

V9FS_PORT
v9fs default port