Trait rs9p::serialize::Encodable

pub trait Encodable {
    fn encode<W: WriteBytesExt>(&self, w: &mut W) -> Result<usize>;
}

Trait representing a type which can be serialized into binary

Required Methods

fn encode<W: WriteBytesExt>(&self, w: &mut W) -> Result<usize>

Encode self to w and returns the number of bytes encoded