Trait rs9p::srv::Filesystem

pub trait Filesystem {
    type Fid;
    fn rstatfs(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall> { ... }

    fn rlopen(&mut self, _: &mut Fid<Self::Fid>, _flags: u32) -> Result<Fcall> { ... }

    fn rlcreate(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _name: &str, 
        _flags: u32, 
        _mode: u32, 
        _gid: u32
    ) -> Result<Fcall> { ... }

    fn rsymlink(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _name: &str, 
        _sym: &str, 
        _gid: u32
    ) -> Result<Fcall> { ... }

    fn rmknod(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _name: &str, 
        _mode: u32, 
        _major: u32, 
        _minor: u32, 
        _gid: u32
    ) -> Result<Fcall> { ... }

    fn rrename(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _: &mut Fid<Self::Fid>, 
        _name: &str
    ) -> Result<Fcall> { ... }

    fn rreadlink(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall> { ... }

    fn rgetattr(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _req_mask: GetattrMask
    ) -> Result<Fcall> { ... }

    fn rsetattr(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _valid: SetattrMask, 
        _stat: &SetAttr
    ) -> Result<Fcall> { ... }

    fn rxattrwalk(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _: &mut Fid<Self::Fid>, 
        _name: &str
    ) -> Result<Fcall> { ... }

    fn rxattrcreate(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _name: &str, 
        _attr_size: u64, 
        _flags: u32
    ) -> Result<Fcall> { ... }

    fn rreaddir(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _offset: u64, 
        _count: u32
    ) -> Result<Fcall> { ... }

    fn rfsync(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall> { ... }

    fn rlock(&mut self, _: &mut Fid<Self::Fid>, _lock: &Flock) -> Result<Fcall> { ... }

    fn rgetlock(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _lock: &Getlock
    ) -> Result<Fcall> { ... }

    fn rlink(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _: &mut Fid<Self::Fid>, 
        _name: &str
    ) -> Result<Fcall> { ... }

    fn rmkdir(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _name: &str, 
        _mode: u32, 
        _gid: u32
    ) -> Result<Fcall> { ... }

    fn rrenameat(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _oldname: &str, 
        _: &mut Fid<Self::Fid>, 
        _newname: &str
    ) -> Result<Fcall> { ... }

    fn runlinkat(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _name: &str, 
        _flags: u32
    ) -> Result<Fcall> { ... }

    fn rauth(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _uname: &str, 
        _aname: &str, 
        _n_uname: u32
    ) -> Result<Fcall> { ... }

    fn rattach(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _afid: Option<&mut Fid<Self::Fid>>, 
        _uname: &str, 
        _aname: &str, 
        _n_uname: u32
    ) -> Result<Fcall> { ... }

    fn rflush(&mut self, _old: Option<&mut Fcall>) -> Result<Fcall> { ... }

    fn rwalk(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _new: &mut Fid<Self::Fid>, 
        _wnames: &[String]
    ) -> Result<Fcall> { ... }

    fn rread(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _offset: u64, 
        _count: u32
    ) -> Result<Fcall> { ... }

    fn rwrite(
        &mut self, 
        _: &mut Fid<Self::Fid>, 
        _offset: u64, 
        _data: &Data
    ) -> Result<Fcall> { ... }

    fn rclunk(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall> { ... }

    fn rremove(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall> { ... }

    fn rversion(&mut self, ms: u32, ver: &str) -> Result<Fcall> { ... }
}

Filesystem server trait.

Implementors can represent an error condition by returning an Err. Otherwise, they must return Fcall with the required fields filled.

The default implementation, returning EOPNOTSUPP error, is provided to the all methods except Rversion. The default implementation of Rversion returns a message accepting 9P2000.L.

NOTE

Defined as Srv in 9p.h of Plan 9.

Protocol

9P2000.L

Associated Types

type Fid

User defined fid type to be associated with a client's fid.

Provided Methods

fn rstatfs(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall>

fn rlopen(&mut self, _: &mut Fid<Self::Fid>, _flags: u32) -> Result<Fcall>

fn rlcreate(
&mut self,
_: &mut Fid<Self::Fid>,
_name: &str,
_flags: u32,
_mode: u32,
_gid: u32
) -> Result<Fcall>

fn rsymlink(
&mut self,
_: &mut Fid<Self::Fid>,
_name: &str,
_sym: &str,
_gid: u32
) -> Result<Fcall>

fn rmknod(
&mut self,
_: &mut Fid<Self::Fid>,
_name: &str,
_mode: u32,
_major: u32,
_minor: u32,
_gid: u32
) -> Result<Fcall>

fn rrename(
&mut self,
_: &mut Fid<Self::Fid>,
_: &mut Fid<Self::Fid>,
_name: &str
) -> Result<Fcall>

fn rreadlink(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall>

fn rgetattr(
&mut self,
_: &mut Fid<Self::Fid>,
_req_mask: GetattrMask
) -> Result<Fcall>

fn rsetattr(
&mut self,
_: &mut Fid<Self::Fid>,
_valid: SetattrMask,
_stat: &SetAttr
) -> Result<Fcall>

fn rxattrwalk(
&mut self,
_: &mut Fid<Self::Fid>,
_: &mut Fid<Self::Fid>,
_name: &str
) -> Result<Fcall>

fn rxattrcreate(
&mut self,
_: &mut Fid<Self::Fid>,
_name: &str,
_attr_size: u64,
_flags: u32
) -> Result<Fcall>

fn rreaddir(
&mut self,
_: &mut Fid<Self::Fid>,
_offset: u64,
_count: u32
) -> Result<Fcall>

fn rfsync(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall>

fn rlock(&mut self, _: &mut Fid<Self::Fid>, _lock: &Flock) -> Result<Fcall>

fn rgetlock(&mut self, _: &mut Fid<Self::Fid>, _lock: &Getlock) -> Result<Fcall>

fn rlink(
&mut self,
_: &mut Fid<Self::Fid>,
_: &mut Fid<Self::Fid>,
_name: &str
) -> Result<Fcall>

fn rmkdir(
&mut self,
_: &mut Fid<Self::Fid>,
_name: &str,
_mode: u32,
_gid: u32
) -> Result<Fcall>

fn rrenameat(
&mut self,
_: &mut Fid<Self::Fid>,
_oldname: &str,
_: &mut Fid<Self::Fid>,
_newname: &str
) -> Result<Fcall>

fn runlinkat(
&mut self,
_: &mut Fid<Self::Fid>,
_name: &str,
_flags: u32
) -> Result<Fcall>

fn rauth(
&mut self,
_: &mut Fid<Self::Fid>,
_uname: &str,
_aname: &str,
_n_uname: u32
) -> Result<Fcall>

fn rattach(
&mut self,
_: &mut Fid<Self::Fid>,
_afid: Option<&mut Fid<Self::Fid>>,
_uname: &str,
_aname: &str,
_n_uname: u32
) -> Result<Fcall>

fn rflush(&mut self, _old: Option<&mut Fcall>) -> Result<Fcall>

fn rwalk(
&mut self,
_: &mut Fid<Self::Fid>,
_new: &mut Fid<Self::Fid>,
_wnames: &[String]
) -> Result<Fcall>

fn rread(
&mut self,
_: &mut Fid<Self::Fid>,
_offset: u64,
_count: u32
) -> Result<Fcall>

fn rwrite(
&mut self,
_: &mut Fid<Self::Fid>,
_offset: u64,
_data: &Data
) -> Result<Fcall>

fn rclunk(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall>

fn rremove(&mut self, _: &mut Fid<Self::Fid>) -> Result<Fcall>

fn rversion(&mut self, ms: u32, ver: &str) -> Result<Fcall>